add app.js
